add route_utils and proper dependency injection
