+++ src/audit/pull_from_queue.py
@@ -6,7 +6,8 @@ import traceback
 from . import logger
 from .config import config
 from .models import CATEGORY_TO_MODEL_CLASS
-from .routes.maintain import insert_row, validate_presigned_url_log, validate_login_log
+from .utils.route_utils import validate_presigned_url_log, validate_login_log
+from .db import get_data_access_layer
 
 
 async def process_log(
@@ -23,14 +24,17 @@ async def process_log(
     if not data.get("timestamp"):
         data["timestamp"] = timestamp
 
-    # validate log
-    if category == "presigned_url":
-        validate_presigned_url_log(data)
-    elif category == "login":
-        validate_login_log(data)
+    async with get_data_access_layer() as dal:
+        # validate log
+        if category == "presigned_url":
+            validate_presigned_url_log(data)
+            dal.create_presigned_url_log(data)
+        elif category == "login":
+            validate_login_log(data)
+            dal.create_login_log(data)
 
     # insert log in DB
-    await insert_row(category, data)
+    # await insert_row(category, data)
 
 
 async def pull_from_queue(sqs):
