+++ src/audit/routes/maintain.py
@@ -7,65 +7,17 @@ from starlette.status import (
 
 from .. import logger
 from ..auth import Auth
+from ..utils.route_utils import (
+    validate_login_log,
+    validate_presigned_url_log,
+)
 from ..db import DataAccessLayer, get_data_access_layer
 from ..models import (
-    CATEGORY_TO_MODEL_CLASS,
     CreateLoginLogInput,
     CreatePresignedUrlLogInput,
 )
 
-
 router = APIRouter()
-
-
-async def insert_row(
-    category, data, dal: DataAccessLayer = Depends(get_data_access_layer)
-):
-    """
-    Insert a new row in the database.
-    """
-    try:
-        if category == "presigned_url":
-            await dal.create_presigned_url_log(data)
-        elif category == "login":
-            await dal.create_login_log(data)
-        else:
-            raise Exception(
-                f"Unknown log category '{category}'",
-            )
-    except Exception:
-        logger.error(
-            f"Failed to insert {category} audit log for URL {data.get('request_url')} at {data.get('timestamp')}"
-        )
-        raise
-
-
-def handle_timestamp(data):
-    """
-    If the timestamp is omitted from the request body, we use the current date
-    and time. In most cases, services should NOT provide a timestamp when
-    creating audit logs. The timestamp is only accepted in log creation
-    requests to allow populating the audit database with historical data, for
-    example by parsing historical logs from before the Audit Service was
-    deployed to a Data Commons.
-    """
-    if "timestamp" not in data:
-        return
-    if data["timestamp"]:
-        # we take a timestamp as input, but store a datetime in the database
-        try:
-            data["timestamp"] = datetime.fromtimestamp(data["timestamp"])
-        except Exception as e:
-            raise HTTPException(
-                HTTP_400_BAD_REQUEST,
-                f"Invalid timestamp '{data['timestamp']}'",
-            )
-    else:
-        # when hitting the API endpoint, the "timestamp" key always exists
-        # because it's defined in `CreateLogInput`. It is automatically added
-        # to rows without timestamp, but we have to remove it from the dict
-        # before inserting in the DB
-        del data["timestamp"]
 
 
 @router.post("/log/presigned_url", status_code=HTTP_201_CREATED)
@@ -73,6 +25,7 @@ async def create_presigned_url_log(
     body: CreatePresignedUrlLogInput,
     background_tasks: BackgroundTasks,
     auth=Depends(Auth),
+    dal: DataAccessLayer = Depends(get_data_access_layer),
 ) -> None:
     """
     Create a new `presigned_url` audit log.
@@ -87,23 +40,15 @@ async def create_presigned_url_log(
     database, so that POSTing audit logs does not impact the performance of
     the caller and audit-service failures are not visible to users.
     """
-    data = body.dict()
+    data = body.model_dump()
     validate_presigned_url_log(data)
-    background_tasks.add_task(insert_row, "presigned_url", data)
-
-
-def validate_presigned_url_log(data):
-    logger.debug(f"Creating `presigned_url` audit log. Received body: {data}")
-
-    allowed_actions = ["download", "upload"]
-    # `action` is a required field", but that's checked during the DB insert
-    if "action" in data and data["action"] not in allowed_actions:
-        raise HTTPException(
-            HTTP_400_BAD_REQUEST,
-            f"Action '{data['action']}' is not allowed ({allowed_actions})",
+    try:
+        await dal.create_presigned_url_log(data)
+    except Exception as e:
+        logger.error(
+            f"Failed to insert presigned_url audit log for URL {data.get('request_url')} at {data.get('timestamp')}"
         )
-
-    handle_timestamp(data)
+        raise
 
 
 @router.post("/log/login", status_code=HTTP_201_CREATED)
@@ -111,6 +56,7 @@ async def create_login_log(
     body: CreateLoginLogInput,
     background_tasks: BackgroundTasks,
     auth=Depends(Auth),
+    dal: DataAccessLayer = Depends(get_data_access_layer),
 ) -> None:
     """
     Create a new `login` audit log.
@@ -125,14 +71,15 @@ async def create_login_log(
     database, so that POSTing audit logs does not impact the performance of
     the caller and audit-service failures are not visible to users.
     """
-    data = body.dict()
+    data = body.model_dump()
     validate_login_log(data)
-    background_tasks.add_task(insert_row, "login", data)
-
-
-def validate_login_log(data):
-    logger.debug(f"Creating `login` audit log. Received body: {data}")
-    handle_timestamp(data)
+    try:
+        await dal.create_login_log(data)
+    except Exception as e:
+        logger.error(
+            f"Failed to insert login audit log for URL {data.get('request_url')} at {data.get('timestamp')}"
+        )
+        raise
 
 
 def init_app(app: FastAPI):
